mtd: spi-nor: introduce spi_nor_read_id()
Follow the upstream linux SPI NOR driver and introduce a dedicated method
to model the read ID operation. As in linux, the newly introduced method
allows callers to pass the number of required dummy bytes, as there are
flashes that need a specific number of dummy cycles for the Read ID
command. No change in functionality intended.

Signed-off-by: Takahiro Kuwano <Takahiro.Kuwano@infineon.com>